Switch to cmake based makefile generation
This change is still in a crude state, but has basic parity with the old
system. The old system had a few drawbacks:
- Many of the targets were only in leaf makefiles. There was a lot of inconsistency,
with some missing targets, and some having incorrect values. The new system allows
generating all executable makefiles using a standard template.
- Dependencies partially worked, but did not check between targets. For example,
if library source code has been changed and an app that uses the library is built,
the system will rebuild the library. I had hacks to do this in specific instances
where it bit me in the old system: for example, to rebuild the bootloader when
synthesizing the design. The new system makes that cleaner and more consistent.
- CMake should make it easier to support on other systems and with other IDEs.

A few notes about this implementation
- This is only supports in-tree builds, which leave more cmake boogers in the tree.
I did it this way because the old system worked like that and I wanted to keep
the changes minimal. I may revisit this in the future, but there are some issues
I need to work through.
- I left the render test makefiles intact. I'm thinking about rewriting this to use
the runtest scripts anyway for consistency with the rest of the tests.  This is
going to take a bit more thought to figure out.
- I haven't added 'Makefile' to .gitignore yet, since there are a few still left around
so the autogenerated Makefiles kind of make a mess of git status.
- Similarly, the generated executables are showing up in git status. Not sure
how to work around that easily.
- I haven't gotten the 'run_debug' script working yet. Need to spend more time
getting escaping working in cmake.
- It might be preferable to generate the run_XXX scripts at cmake configure time
instead of during compile time.

+++ software/libs/librender/RenderContext.cpp
@@ -97,7 +97,7 @@ void RenderContext::_wireframeTile(void *_castToContext, int index)
 
 void RenderContext::finish()
 {
-    int kMaxTiles = fTileColumns * fTileRows;
+    unsigned int kMaxTiles = static_cast<unsigned int>(fTileColumns * fTileRows);
     fTiles = new (fAllocator) TriangleArray[kMaxTiles];
     for (int i = 0; i < kMaxTiles; i++)
         fTiles[i].setAllocator(&fAllocator);
